move shared validation functions in utils module
